BIG WARNING HERE!
The trick with an empty virtual destructor at Hander was an attempt to avoid the implicit delete call (jump actually) generated by GCC. It wasn't working smoothly, so modifications were made to have a in-library version of memory allocation functions (i.e. delete), so that now GCC can issue calls to them.

This, however, might be just another silly workaround, since we still miss the explanation of why GCC is doing that, and, perhaps, will end up deleting something we didn't allocated!