Due to a bug in setup and also in the direct deployment of mediators by applications, libsystem.a was split in three: - libarch.a:  Processor (CPU, MMU, TSC) - libmach.a: Machine (Bus, IC, etc) - libsystem.a: system abstractions
Methods init() in ARCH and MACH are still at libinit, with was unchanged. The rule of thumb here is that libarch.a, libmach.a and libinit.a MUST NOT contain global constructors.

The bug? Having a global constructor in a ".o" causes it to be linked independently of being used. SETUP and APPLICATION were getting almost whole copies of SYSTEM because of Sytem::_heap;