A major redesign was carried out aiming at isolating the scheduler and thus easing the implementation of new scheduling disciplines (also in hardware).
Scheduler is now an object, which maintain any associated queues. The former priority was replaced by a more generic "scheduling criterion" to be selected at traits.h. This object might (or not) use the scheduling list utility class.

As regards real-time, rt_thread was removed from the repository for a while, since it should collaborate with other classes and go deeper into RT questions. A periodic_thread.h was created in its place with the sense that the new scheduler is able to implement policies for both: aperiodic and periodic threads.