BIG FAT WARNING HERE!
The reeschedule functions are using intermediate variables, otherwise the system would crash. This makes no sense! There must be a hidden bug around the volatility of chosen!

---

A major redesign was carried out aiming at isolating the scheduler and thus easing the implementation of new scheduling disciplines (also in hardware).

Scheduler is now an object, which maintain any associated queues. The former priority was replaced by a more generic "scheduling criterion" to be selected at traits.h. This object might (or not) use the scheduling list utility class.