Display was long being subject of discussion because to avoid remapping of Frame Buffer, it had a constructor that reset the pointer before every use. This was painful for Serial_Display, which had to reconfigure UART before every access.
With SMP, this problem became more evident and a redesign was make so the Frame Buffer pointer is now a class attribute. This allowed methods to go "static" and Display to be used without an object, but introduces the trouble with remapping the pointer whenever the memory model is changed (e.g. from SETUP to INIT to APP).