Correction in grouping list used by the HEAP: Grouping_List. The HEAP uses Grouping_List for control of "unallocated spaces".
Cause:
Grouping_List allowed unallocated spaces smaller than the sizeof(Element).
Each unallocated space is represented by an Element object.
And each Element has the same address of the unallocated space, as well as its size.
So when an unallocated space obtained size < sizeof(Element), the Element object itself was corrupted.
It caused out of memory or the EPOS was restarted (depending on the corrupted values).

Correction:
    - Changes in the method Grouping_List::search_size(unsigned int).
         - Now, the method looks for elements that the leftover of unallocated spaces on the heap is
           greater than or equal to sizeof (Element), or that the leftover is exactly equal to 0.