BIG WARNING here!
The solution to make EPOS independent of a particular GCC version implied in establishing .ctros (the global constructor table) as the first section in the DATA segment. This was not true for all GCC versions: some moved that table, along with .dtors and .eh_frame to the CODE segment. This was very good for MCUs with little memory.

This must be further investigated!