Interrupt_Id passed by const reference
The idea of using const Type & for typedefs to support eventual future
replacement for a complex type (e.g. a struct) was very bad in this
case, since a lot of methods receiving Interrupt_Id are being defined
outside the classes and therefore a true reference is being passed in
software (i.e. a pointer), but just and id when done in hardware.