Priority Ceiling
This is a new priority ceiling scheme. Instead of raising the priority
of the running thread and calling reschedule() to let the Scheduler know
about it, it simply changes the rank (i.e. the priority) of the element
(always head) in the scheduling list. The previous strategy was not
passing many tests and was causing a lot of jitter. This one must be
carefully observed.