Machine::pre_init() and CPU::smp_barrier_init()
For machines that enable multicore and wake up all cores during SETUP
(should be all, isn't it?), it is impossible to tell which core will get
to Machine::pre_init() first, so conditioning the initialization of SMP
barrier and Display to the BSP doesn't make sense. On the other hand,
machines that do not enable multicore (e.g. Cortex-A9; why?) need a
guard here to prevent multiple invocations of wake up. I opted for
replacing CPU::smp_init() by CPU::smp_barrier_init() to make it clear
that this is all what should be happening here (barrier initialization).