Fix GPIO
The PL061 setter must grant the value inserted is a multiple of 0xff. GPIO class receives a boolean that is forwarded to this method, thus being necessary PL061 setter to multiply the received value per 0xff, or to adjust the components that use this engine to do it before to pass the char value to be written.
There was also a change on emote3_usb to fix the start of the component.
Note: Need to take a look to the one-parameter set method on pl061.h. It is not considering the shift presented on manual.