Transducer relation with Connection Devices
Transducer often use I2C, SPI, UART or a similar connection. The
previous design defined them to be static so they could be reused by
multi-transducers (like IMU). However, it suffered from the know problem
of relying on global initialization ordering or on requiring the
attribute corresponding to the connection device to be initialized
before the base class (the transducer engine itself).

This new design takes on the fact that C++ allows aggregation of
cardinality 1 to be implemented as inheritance and solves the problem by
making Transducers to inherit first from Connection and later from the
Engine.

The drawback is that multi-transducers will initialize the connection
every time one of its facets are used (e.g. new Accelerometer
initialized I2C, and new Gyroscope initializes it again).