Adding Flag -fno-pie
Code on Ubuntu and Debian was compiling but not running on qemu!
Analyzing the objdump we could find the __x86.get_pc_thunk.bx on .init section. A stack overflow forum suggests adding the flag -fno-pie to get rid of the calls for this address.
According to gcc man-page, no-pie flags tells the compiler to: Don't produce a position independent executable.
Link to stack forum: https://stackoverflow.com/questions/50105581/how-do-i-get-rid-of-call-x86-get-pc-thunk-ax .
Adding -fno-pie to MACH_CXX_FLAGS was enough to make code run on qemu, however, adding it to MACH_CC_FLAGS too makes it compiles generating the same _start address as Fedora compiler and our pre-compiled gcc-7.2.0.