Fix SysInfo for eMote3, Make code to read Extras on Raspberry_Pi
Raspberry Pi extras can not be linked with mkbi since objcopy removes what is not code. Thus, the extras are added now via cat command.
Raspberry Pi presented a bug when reading a signature with more than 10 characteres, therefore it was chosen a different signature from Intel.

eMote3 changed a bit.
the hex conversion is now done with the .img file as source.
To support the sysinfo linking, the mkbi writes sysinfo on .img file and do not writes machine secrets on the .img for eMote3 (so the conversion from .img to .hex is possible). When mkbi works with hex file, it does not write sysinfo (it is already on the file) and writes machine secrets.