BugFix: Float Point Operations Raspberry ANN
After a multitude of tests, where ANN results were unconsistent with training results,
The FPU usage seems to be the one responsible for such problems. Added push and pop of FPU
registers when entering and exiting ANN code to preserve correctness.
For the first test case, it seems to be working fine now