fixing I2C get and send
when the I2C data trasmission is complete, the busy bit of I2C stat is cleared.Therefore, before returning read data or confirming a trasmission acknowledgment, a loop waits for the busy bit to be cleared (CC2538 User's Guide, Version C, Texas Instruments, April 2012). However it was shown necessary that we have sure the busy bit has been written before we enter the loop that waits for it to be cleared.
Also, the wait_to_sync parameter on USB traits is set to true by default, it is only supposed to be false when measuring performance.