Returning some of the clobbers
According to the documentation of gcc (https://gcc.gnu.org/onlinedocs/gcc/Extended-Asm.html#Clobbers-and-Scratch-Registers), a register must be in the clobbered list whenever we use it and it is not an output or an input for the asm(). Clobbering the SP, from the compiler's perspective, is invalid. We know we have to do it, but to avoid warnings from the compiler, it was removed.