Multicore
A major redesign of locking: now, no implicit unlocks are expected. Dispatch unlocks before context switch and re-acquires the lock before proceeding with the next thread. It passed the main test on x86, but it is a big change that must be thoroughly validated. Some migrations were performed by Leonardo with the new priority().