* SmartData
Time management at the API is broken and must be replaced/agreed with Microsecond. The next step will be to redesign TSTP and SmartData to work with Microseconds since EPOCH and let the dirty, hard work of compressing it to the MAC. Buffer::metadata() is broken, but I don't see why.

Many data structures dependent from the KEY_SIZE of the Cipher were not properly bound.

TSTP::Router::destination() was considering Expiry an absolute value. I  guess it is relative to t0, isn't it?

Timekeeper::keep_alive was setting TR on replies, which would complicate the router while deciding to route or not. I removed it, so TR is only set for the original requesting node.

TSTP::Timekeeper::Timekeeper() in time_init.cc had a concurring handler and loop, both sending TRs. I couldn't understand it and it wasn't working. I removed it.

* Monitor

PMU::_events constexpr is not working on new compiles. Disabling monitor temporarily.

Adding heap::free() to reserve memory for Monitor at initialization.

Removing PMU::read() calls for QEMU tests

Reading PMU while in QEMU is only available with KVM enabled, thus, if this feature is not available, the execution will stop at the first PMU::read(). The tests are functional when running the applications in a real machine (or enabling the KVM feature).

*  NIC

Time_Stamp moved to Statistics (with no extra overhead since only that field is updated on the calls to statistics() and a reference to the object is returned. Each NIC_Family defines the resolution, with eventual conversions handled by the protocols.

 There was also a bug in handling accuracy: TSC uses PPB, while the NIC uses PPM!

* BUG FIX: Monitored on Thread Creation

The _task thread list was being accessed when the system was configured to monitored = true even on single-task application, thus,  breaking the application execution during the creation of the first thread.

* BUGFIX: IA32::clock()
The current clock was not initialized.

*  BUGFIX: PC_Timer

Timer::reset() used to call Engine::config() and this was replaced by a hard reset through Engine::reset(), but it has the impact of slowing down time a lot if i8253 is selected. Actually, there was a comment about the absence of a reset for i8253. So I rolled back to Engine::config().

* eMote3: many bug fixes by Roberto Scheffer, including ADC, GPIO, and I2C. cleanup