constexpr
According to the C++ standard, constexpr does not need to be taken as a constant expression by the compiler: it means the compiler must try. Previous versions of GCC used to optimize the use of constexpr arrays in functions that are themselves not constexpr, but the newer versions require all client functions to be also constexpr. For a while, the solution was to move the arrays to separate units (.cc). It seems they are not always ending at the final executable, but the more robust solution would make many other functions that use them constexpr too.