BUGFIX: Built-in Architecture
In BUILTIN, EPOS was taking the entry point of the main thread from the
same __epos_app_entry using in LIBRARY. This worked in the past because
hardware mediators used to have empty constructors, so calling _init
twice, once for the system and another for the application, was
harmless. Now, with non-empty constructors, BUILTIN must use the same
strategy of KERNEL. That is, get the entry point from Sysmtem_Info.